+++ ds/org.eclipse.pde.ds.annotations.tests/META-INF/MANIFEST.MF
@@ -2,20 +2,22 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: DS Annotations Tests
 Bundle-SymbolicName: org.eclipse.pde.ds.annotations.tests
-Bundle-Version: 1.4.0.qualifier
+Bundle-Version: 1.4.100.qualifier
 Bundle-Activator: org.eclipse.pde.ds.internal.annotations.tests.Activator
 Bundle-Vendor: Eclipse.org
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
-Require-Bundle: org.eclipse.pde.ds.annotations;bundle-version="[1.1.0,1.5.0)",
- org.eclipse.pde.ds.core;bundle-version="[1.1.0,2.0.0)",
- org.eclipse.pde.ui;bundle-version="[3.9.0,4.0.0)",
- org.eclipse.core.resources;bundle-version="[3.11.0,4.0.0)",
- org.eclipse.core.runtime;bundle-version="[3.29.0,4.0.0)",
- org.eclipse.text;bundle-version="[3.6.0,4.0.0)"
+Require-Bundle: org.eclipse.pde.ds.annotations;bundle-version="[1.4.0,1.5.0)",
+ org.eclipse.pde.ds.core;bundle-version="[1.3.0,2.0.0)",
+ org.eclipse.pde.ui;bundle-version="[3.16.0,4.0.0)",
+ org.eclipse.core.resources;bundle-version="[3.23.0,4.0.0)",
+ org.eclipse.core.runtime;bundle-version="[3.34.0,4.0.0)",
+ org.eclipse.text;bundle-version="[3.14.0,4.0.0)"
 Export-Package: org.eclipse.pde.ds.internal.annotations.tests;x-internal:=true
 Import-Package: org.junit.jupiter.api;version="[5.14.0,6.0.0)",
- org.junit.platform.suite.api;version="[1.14.0,2.0.0)"
+ org.junit.jupiter.api.extension;version="[5.14.0,6.0.0)",
+ org.junit.platform.suite.api;version="[1.14.0,2.0.0)",
+ org.opentest4j;version="[1.3.0,2.0.0)"
 Eclipse-BundleShape: dir
 Bundle-ClassPath: tests.jar
 Automatic-Module-Name: org.eclipse.pde.ds.annotations.tests
